requirements.txt
django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-extensions==3.2.3
ipython==9.0.2
